+++ CHANGES.md
@@ -1,5 +1,6 @@
 
 
+- Build depend on topkg.
 - Relicense from BSD3 to ISC.
 
 v1.0.0 2015-06-17 Cambridge (UK)